+++ tests/test_ml_streaming/test_streaming.py
@@ -13,6 +13,17 @@ from scenarios.scenario3 import scenario3
     params=["NULL", "GZIP", "ZLIB", "LZMA", "BZ2"])
 def prepare_bare_server(request):
     compression_method = request.param
+    # Using PyMilo direct CLI
+    # server_proc = subprocess.Popen(
+    #     [
+    #         executable,
+    #         "-m", "pymilo",
+    #         "--compression", compression_method,
+    #         "--protocol", "REST",
+    #         "--port", "8000",
+    #         "--bare",
+    #     ],
+    #     )
     path = os.path.join(
         os.getcwd(),
         "tests",
@@ -28,7 +39,7 @@ def prepare_bare_server(request):
         ],
         )
     time.sleep(10)
-    yield (server_proc, compression_method, "REST")
+    yield (compression_method, "REST")
     server_proc.terminate()
 
 
@@ -38,7 +49,18 @@ def prepare_bare_server(request):
 def prepare_ml_server(request):
     communication_protocol = request.param
     compression_method = "ZLIB"
-    print(communication_protocol)
+    # Using PyMilo direct CLI
+    # server_proc = subprocess.Popen(
+    #     [
+    #         executable,
+    #         "-m", "pymilo",
+    #         "--compression", compression_method,
+    #         "--protocol", communication_protocol,
+    #         "--port", "9000",
+    #         "--load", os.path.join(os.getcwd(), "tests", "test_exceptions", "valid_jsons", "linear_regression.json")
+    #         # "--load", "https://raw.githubusercontent.com/openscilab/pymilo/main/tests/test_exceptions/valid_jsons/linear_regression.json",
+    #     ],
+    #     )
     path = os.path.join(
         os.getcwd(),
         "tests",
@@ -54,21 +76,21 @@ def prepare_ml_server(request):
             "--init",
         ],
         )
-    time.sleep(5)
-    yield (server_proc, compression_method, communication_protocol)
+    time.sleep(10)
+    yield (compression_method, communication_protocol)
     server_proc.terminate()
 
 
 def test1(prepare_bare_server):
-    _, compression_method, communication_protocol = prepare_bare_server
+    compression_method, communication_protocol = prepare_bare_server
     assert scenario1(compression_method, communication_protocol) == 0
 
 
 def test2(prepare_bare_server):
-    _, compression_method, communication_protocol = prepare_bare_server
+    compression_method, communication_protocol = prepare_bare_server
     assert scenario2(compression_method, communication_protocol) == 0
 
 
 def test3(prepare_ml_server):
-    _, compression_method, communication_protocol = prepare_ml_server
+    compression_method, communication_protocol = prepare_ml_server
     assert scenario3(compression_method, communication_protocol) == 0